+++ resources/views/bootstrap-4/includes/table.blade.php
@@ -39,6 +39,7 @@
                     <x-livewire-tables::bs4.table.cell>
                         <input
                             wire:model="selected"
+                            wire:loading.attr.delay="disabled"
                             value="{{ $row->{$primaryKey} }}"
                             onclick="event.stopPropagation();return true;"
                             type="checkbox"