+++ resources/views/tailwind/includes/bulk-select-row.blade.php
@@ -11,6 +11,7 @@
 
                     <button
                         wire:click="resetBulk"
+                        wire:loading.attr="disabled"
                         type="button"
                         class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
                     >
@@ -27,6 +28,7 @@
 
                     <button
                         wire:click="resetBulk"
+                        wire:loading.attr="disabled"
                         type="button"
                         class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
                     >
@@ -44,6 +46,7 @@
 
                         <button
                             wire:click="resetBulk"
+                            wire:loading.attr="disabled"
                             type="button"
                             class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
                         >
@@ -61,6 +64,7 @@
 
                         <button
                             wire:click="selectAll"
+                            wire:loading.attr="disabled"
                             type="button"
                             class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
                         >
@@ -69,6 +73,7 @@
 
                         <button
                             wire:click="resetBulk"
+                            wire:loading.attr="disabled"
                             type="button"
                             class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
                         >