+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 ### Added
 
 - Added Unselect All button on bulk row when selecting page.
+- Added disabled delay on select checkboxes.
+- Added disabled on bulk row button clicks.
 
 ### Changed